# demo/src/main/java/com/example/demo/dto/UserDTO.java
package com.example.demo.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.Setter;

@AllArgsConstructor
public class UserDTO {
    public String username;
    public String password;
    public String email;

}